✨api: fix tests
