+++ api/tests/certification/results/acceptance/application/livret-scolaire-route_test.js
@@ -1,6 +1,6 @@
 import { Assessment } from '../../../../../src/shared/domain/models/index.js';
 import {
-  createServer,
+  createMaddoServer,
   databaseBuilder,
   expect,
   generateAuthenticatedUserRequestHeaders,
@@ -202,7 +202,7 @@ describe('Certification | Results | Acceptance | Application | Livret Scolaire',
     context('when the given uai is correct', function () {
       it('should return 200 HTTP status code with the certifications results and referential of competences', async function () {
         // given
-        server = await createServer();
+        server = await createMaddoServer();
         const user = buildUser();
         const organizationLearner = buildOrganizationLearner({ userId: user.id, organizationId });
         const { session, certificationCourse } = buildValidatedPublishedCertificationData({
@@ -281,7 +281,7 @@ describe('Certification | Results | Acceptance | Application | Livret Scolaire',
     context('when the given uai is incorrect', function () {
       it('should return 200 HTTP status code with the referential of competences only', async function () {
         // given
-        server = await createServer();
+        server = await createMaddoServer();
         options = {
           method: 'GET',
           url: '/api/organizations/9999/certifications',
@@ -319,7 +319,7 @@ describe('Certification | Results | Acceptance | Application | Livret Scolaire',
     context('when the access token is generated by password credential grant to authenticate end user', function () {
       it('should return unauthorized status code', async function () {
         // given
-        server = await createServer();
+        server = await createMaddoServer();
         options = {
           method: 'GET',
           url: '/api/organizations/9999/certifications',
@@ -333,46 +333,5 @@ describe('Certification | Results | Acceptance | Application | Livret Scolaire',
         expect(response.statusCode).to.equal(401);
       });
     });
-
-    context('when the access token is generated by client credential grant with the wrong client id', function () {
-      it('should return unauthorized status code', async function () {
-        // given
-        server = await createServer();
-        options = {
-          method: 'GET',
-          url: '/api/organizations/9999/certifications',
-          headers: { authorization: generateValidRequestAuthorizationHeaderForApplication('') },
-        };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(401);
-      });
-    });
-
-    context(
-      'when the access token is generated by client credential grant with wrong livret scolaire scope',
-      function () {
-        it('should return Forbidden access status code', async function () {
-          // given
-          server = await createServer();
-          options = {
-            method: 'GET',
-            url: '/api/organizations/9999/certifications',
-            headers: {
-              authorization: generateValidRequestAuthorizationHeaderForApplication(OSMOSE_CLIENT_ID, OSMOSE_SOURCE),
-            },
-          };
-
-          // when
-          const response = await server.inject(options);
-
-          // then
-          expect(response.statusCode).to.equal(403);
-        });
-      },
-    );
   });
 });
