+++ api/tests/certification/results/unit/application/livret-scolaire-route_test.js
@@ -1,0 +1,67 @@
+import * as moduleUnderTest from '../../../../../src/certification/results/application/livret-scolaire-route.js';
+import {
+  expect,
+  generateValidRequestAuthorizationHeaderForApplication,
+  HttpTestServer,
+} from '../../../../test-helper.js';
+
+describe('Certification | Results | Unit | Application | Livret Scolaire', function () {
+  let headers, httpTestServer;
+  const OSMOSE_CLIENT_ID = 'test-apimOsmoseClientId';
+  const OSMOSE_SOURCE = 'livretScolaire';
+
+  describe('GET /api/organizations/{uai}/certifications', function () {
+    context('when the access token is generated by client credential grant with the wrong client id', function () {
+      it('should return unauthorized status code', async function () {
+        // given
+        httpTestServer = new HttpTestServer();
+        httpTestServer.setupAuthentication();
+        await httpTestServer.register(moduleUnderTest);
+
+        headers = {
+          authorization: generateValidRequestAuthorizationHeaderForApplication(''),
+        };
+
+        // when
+        const response = await httpTestServer.request(
+          'GET',
+          '/api/organizations/999/certifications',
+          {},
+          null,
+          headers,
+        );
+
+        // then
+        expect(response.statusCode).to.equal(401);
+      });
+    });
+
+    context(
+      'when the access token is generated by client credential grant with wrong livret scolaire scope',
+      function () {
+        it('should return Forbidden access status code', async function () {
+          // given
+          httpTestServer = new HttpTestServer();
+          httpTestServer.setupAuthentication();
+          await httpTestServer.register(moduleUnderTest);
+
+          headers = {
+            authorization: generateValidRequestAuthorizationHeaderForApplication(OSMOSE_CLIENT_ID, OSMOSE_SOURCE),
+          };
+
+          // when
+          const response = await httpTestServer.request(
+            'GET',
+            '/api/organizations/999/certifications',
+            {},
+            null,
+            headers,
+          );
+
+          // then
+          expect(response.statusCode).to.equal(403);
+        });
+      },
+    );
+  });
+});
